+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/Transaction.java
@@ -766,6 +766,7 @@ public class Transaction
       case FRONTIER:
         preimage = frontierPreimage(nonce, gasPrice, gasLimit, to, value, payload, chainId);
         break;
+      case BLOB: // ToDo 4844: specialize for blob when more field will be added for it
       case EIP1559:
       case BLOB:
         preimage =