+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/messages/MessageWrapperTest.java
@@ -329,6 +329,7 @@ public class MessageWrapperTest {
           Hash.fromHexString(mixHash),
           Bytes.fromHexStringLenient(nonce).toLong(),
           null,
+          null,
           new MainnetBlockHeaderFunctions());
     }
   }